Move normalization to submodule
The semantics module is starting to get quite large, so moving this to a submodule might make things a little easier!
