+++ src/semantics/mod.rs
@@ -8,8 +8,7 @@ use codespan::ByteSpan;
 use moniker::{Binder, BoundTerm, Embed, FreeVar, Scope, Var};
 
 use syntax::core::{
-    Head, Item, Literal, Module, Neutral, Pattern, RcNeutral, RcPattern, RcTerm, RcType, RcValue,
-    Term, Type, Value,
+    Item, Literal, Module, Pattern, RcPattern, RcTerm, RcType, RcValue, Term, Type, Value,
 };
 use syntax::raw;
 use syntax::translation::Resugar;
@@ -17,12 +16,14 @@ use syntax::Level;
 
 mod env;
 mod errors;
+mod normalize;
 mod prim;
 #[cfg(test)]
 mod tests;
 
 pub use self::env::TcEnv;
 pub use self::errors::{InternalError, TypeError};
+pub use self::normalize::{match_value, nf_term};
 pub use self::prim::{PrimEnv, PrimFn};
 
 /// Type check and elaborate a module
@@ -98,7 +99,7 @@ pub fn check_module(tc_env: &TcEnv, raw_module: &raw::Module) -> Result<Module, 
                     // We found a prior declaration, so lets use it as a basis
                     // for checking the definition
                     Some(ForwardDecl::Pending(_, ty)) => {
-                        let ty = normalize(&tc_env, &ty)?;
+                        let ty = nf_term(&tc_env, &ty)?;
                         (check_term(&tc_env, &raw_term, &ty)?, ty)
                     },
                     // No prior declaration was found, so try to infer the type
@@ -121,222 +122,6 @@ pub fn check_module(tc_env: &TcEnv, raw_module: &raw::Module) -> Result<Module, 
     }
 
     Ok(Module { items })
-}
-
-/// Reduce a term to its normal form
-pub fn normalize(tc_env: &TcEnv, term: &RcTerm) -> Result<RcValue, InternalError> {
-    match *term.inner {
-        // E-ANN
-        Term::Ann(ref expr, _) => normalize(tc_env, expr),
-
-        // E-TYPE
-        Term::Universe(level) => Ok(RcValue::from(Value::Universe(level))),
-
-        Term::Literal(ref lit) => Ok(RcValue::from(Value::Literal(lit.clone()))),
-
-        // E-VAR, E-VAR-DEF
-        Term::Var(ref var) => match *var {
-            Var::Free(ref name) => match tc_env.definitions.get(name) {
-                Some(term) => normalize(tc_env, term),
-                None => Ok(RcValue::from(Value::from(var.clone()))),
-            },
-
-            // We should always be substituting bound variables with fresh
-            // variables when entering scopes using `unbind`, so if we've
-            // encountered one here this is definitely a bug!
-            Var::Bound(_) => Err(InternalError::UnexpectedBoundVar {
-                span: None,
-                var: var.clone(),
-            }),
-        },
-
-        Term::Extern(ref name, ref ty) => Ok(RcValue::from(Value::from(Neutral::Head(
-            Head::Extern(name.clone(), normalize(tc_env, ty)?),
-        )))),
-
-        Term::Global(ref name) => match tc_env.globals.get(name.as_str()) {
-            Some(&(Some(ref value), _)) => Ok(value.clone()),
-            Some(&(None, _)) | None => Ok(RcValue::from(Value::global(name.clone()))),
-        },
-
-        // E-PI
-        Term::Pi(ref scope) => {
-            let ((name, Embed(ann)), body) = scope.clone().unbind();
-
-            Ok(RcValue::from(Value::Pi(Scope::new(
-                (name, Embed(normalize(tc_env, &ann)?)),
-                normalize(tc_env, &body)?,
-            ))))
-        },
-
-        // E-LAM
-        Term::Lam(ref scope) => {
-            let ((name, Embed(ann)), body) = scope.clone().unbind();
-
-            Ok(RcValue::from(Value::Lam(Scope::new(
-                (name, Embed(normalize(tc_env, &ann)?)),
-                normalize(tc_env, &body)?,
-            ))))
-        },
-
-        // E-APP
-        Term::App(ref head, ref arg) => {
-            match *normalize(tc_env, head)?.inner {
-                Value::Lam(ref scope) => {
-                    // FIXME: do a local unbind here
-                    let ((Binder(free_var), Embed(_)), body) = scope.clone().unbind();
-                    normalize(tc_env, &body.substs(&[(free_var, arg.clone())]))
-                },
-                Value::Neutral(ref neutral, ref spine) => {
-                    let arg = normalize(tc_env, arg)?;
-                    let mut spine = spine.clone();
-
-                    match *neutral.inner {
-                        Neutral::Head(Head::Extern(ref name, _)) => {
-                            spine.push_back(arg);
-
-                            // Apply the arguments to primitive definitions if the number of
-                            // arguments matches the arity of the primitive, all aof the arguments
-                            // are fully normalized
-                            if let Some(prim) = tc_env.primitives.get(name) {
-                                if prim.arity == spine.len() && spine.iter().all(|arg| arg.is_nf())
-                                {
-                                    match (prim.interpretation)(spine) {
-                                        Ok(value) => return Ok(value),
-                                        Err(()) => unimplemented!("proper error"),
-                                    }
-                                }
-                            }
-                        },
-                        Neutral::Head(Head::Var(_))
-                        | Neutral::Head(Head::Global(_))
-                        | Neutral::If(_, _, _)
-                        | Neutral::Proj(_, _)
-                        | Neutral::Case(_, _) => spine.push_back(arg),
-                    }
-
-                    Ok(RcValue::from(Value::Neutral(neutral.clone(), spine)))
-                },
-                _ => Err(InternalError::ArgumentAppliedToNonFunction),
-            }
-        },
-
-        // E-IF, E-IF-TRUE, E-IF-FALSE
-        Term::If(ref cond, ref if_true, ref if_false) => {
-            let value_cond = normalize(tc_env, cond)?;
-
-            match *value_cond {
-                Value::Literal(Literal::Bool(true)) => normalize(tc_env, if_true),
-                Value::Literal(Literal::Bool(false)) => normalize(tc_env, if_false),
-                Value::Neutral(ref cond, ref spine) => Ok(RcValue::from(Value::Neutral(
-                    RcNeutral::from(Neutral::If(
-                        cond.clone(),
-                        normalize(tc_env, if_true)?,
-                        normalize(tc_env, if_false)?,
-                    )),
-                    spine.clone(),
-                ))),
-                _ => Err(InternalError::ExpectedBoolExpr),
-            }
-        },
-
-        // E-RECORD-TYPE
-        Term::RecordType(ref scope) => {
-            let ((label, binder, Embed(ann)), body) = scope.clone().unbind();
-            let ann = normalize(tc_env, &ann)?;
-            let body = normalize(tc_env, &body)?;
-
-            Ok(Value::RecordType(Scope::new((label, binder, Embed(ann)), body)).into())
-        },
-
-        // E-EMPTY-RECORD-TYPE
-        Term::RecordTypeEmpty => Ok(RcValue::from(Value::RecordTypeEmpty)),
-
-        // E-RECORD
-        Term::Record(ref scope) => {
-            let ((label, binder, Embed(term)), body) = scope.clone().unbind();
-            let value = normalize(tc_env, &term)?;
-            let body = normalize(tc_env, &body)?;
-
-            Ok(Value::Record(Scope::new((label, binder, Embed(value)), body)).into())
-        },
-
-        // E-EMPTY-RECORD
-        Term::RecordEmpty => Ok(RcValue::from(Value::RecordEmpty)),
-
-        // E-PROJ
-        Term::Proj(ref expr, ref label) => match *normalize(tc_env, expr)? {
-            Value::Neutral(ref neutral, ref spine) => Ok(RcValue::from(Value::Neutral(
-                RcNeutral::from(Neutral::Proj(neutral.clone(), label.clone())),
-                spine.clone(),
-            ))),
-            ref expr => match expr.lookup_record(label) {
-                Some(value) => Ok(value.clone()),
-                None => Err(InternalError::ProjectedOnNonExistentField {
-                    label: label.clone(),
-                }),
-            },
-        },
-
-        // E-CASE
-        Term::Case(ref head, ref clauses) => {
-            let head = normalize(tc_env, head)?;
-
-            if let Value::Neutral(ref neutral, ref spine) = *head {
-                Ok(RcValue::from(Value::Neutral(
-                    RcNeutral::from(Neutral::Case(
-                        neutral.clone(),
-                        clauses
-                            .iter()
-                            .map(|clause| {
-                                let (pattern, body) = clause.clone().unbind();
-                                Ok(Scope::new(pattern, normalize(tc_env, &body)?))
-                            }).collect::<Result<_, _>>()?,
-                    )),
-                    spine.clone(),
-                )))
-            } else {
-                for clause in clauses {
-                    let (pattern, body) = clause.clone().unbind();
-                    if let Some(mappings) = match_value(&pattern, &head) {
-                        let mappings = mappings
-                            .into_iter()
-                            .map(|(free_var, value)| (free_var, RcTerm::from(&*value.inner)))
-                            .collect::<Vec<_>>();
-                        return normalize(tc_env, &body.substs(&mappings));
-                    }
-                }
-                Err(InternalError::NoPatternsApplicable)
-            }
-        },
-
-        // E-ARRAY
-        Term::Array(ref elems) => Ok(RcValue::from(Value::Array(
-            elems
-                .iter()
-                .map(|elem| normalize(tc_env, elem))
-                .collect::<Result<_, _>>()?,
-        ))),
-    }
-}
-
-/// If the pattern matches the value, this function returns the substitutions
-/// needed to apply the pattern to some body expression
-pub fn match_value(
-    pattern: &RcPattern,
-    value: &RcValue,
-) -> Option<Vec<(FreeVar<String>, RcValue)>> {
-    match (&*pattern.inner, &*value.inner) {
-        (&Pattern::Literal(ref pattern_lit), &Value::Literal(ref value_lit))
-            if pattern_lit == value_lit =>
-        {
-            Some(vec![])
-        },
-        (&Pattern::Binder(Binder(ref free_var)), _) => {
-            Some(vec![(free_var.clone(), value.clone())])
-        },
-        (_, _) => None,
-    }
 }
 
 /// Ensures that the given term is a universe, returning the level of that
@@ -448,7 +233,7 @@ pub fn infer_pattern(
     match *raw_pattern.inner {
         raw::Pattern::Ann(ref raw_pattern, Embed(ref raw_ty)) => {
             let (ty, _) = infer_universe(tc_env, raw_ty)?;
-            let value_ty = normalize(tc_env, &ty)?;
+            let value_ty = nf_term(tc_env, &ty)?;
             let (pattern, declarations) = check_pattern(tc_env, raw_pattern, &value_ty)?;
 
             Ok((
@@ -530,7 +315,7 @@ pub fn check_term(
 
             if label == ty_label {
                 let expr = check_term(tc_env, &raw_expr, &ann)?;
-                let ty_body = normalize(tc_env, &ty_body.substs(&[(ty_binder.0, expr.clone())]))?;
+                let ty_body = nf_term(tc_env, &ty_body.substs(&[(ty_binder.0, expr.clone())]))?;
                 let body = check_term(tc_env, &raw_body, &ty_body)?;
 
                 return Ok(RcTerm::from(Term::Record(Scope::new(
@@ -620,7 +405,7 @@ pub fn infer_term(tc_env: &TcEnv, raw_term: &raw::RcTerm) -> Result<(RcTerm, RcT
         //  I-ANN
         raw::Term::Ann(ref raw_expr, ref raw_ty) => {
             let (ty, _) = infer_universe(tc_env, raw_ty)?;
-            let value_ty = normalize(tc_env, &ty)?;
+            let value_ty = nf_term(tc_env, &ty)?;
             let expr = check_term(tc_env, raw_expr, &value_ty)?;
 
             Ok((RcTerm::from(Term::Ann(expr, ty)), value_ty))
@@ -670,7 +455,7 @@ pub fn infer_term(tc_env: &TcEnv, raw_term: &raw::RcTerm) -> Result<(RcTerm, RcT
 
         raw::Term::Extern(_, _, ref name, ref raw_ty) => {
             let (ty, _) = infer_universe(tc_env, raw_ty)?;
-            let value_ty = normalize(tc_env, &ty)?;
+            let value_ty = nf_term(tc_env, &ty)?;
             Ok((RcTerm::from(Term::Extern(name.clone(), ty)), value_ty))
         },
 
@@ -688,7 +473,7 @@ pub fn infer_term(tc_env: &TcEnv, raw_term: &raw::RcTerm) -> Result<(RcTerm, RcT
 
             let (ann, ann_level) = infer_universe(tc_env, &raw_ann)?;
             let (body, body_level) = {
-                let ann = normalize(tc_env, &ann)?;
+                let ann = nf_term(tc_env, &ann)?;
                 let mut body_tc_env = tc_env.clone();
                 body_tc_env.declarations.insert(free_var.clone(), ann);
                 infer_universe(&body_tc_env, &raw_body)?
@@ -714,7 +499,7 @@ pub fn infer_term(tc_env: &TcEnv, raw_term: &raw::RcTerm) -> Result<(RcTerm, RcT
             }
 
             let (lam_ann, _) = infer_universe(tc_env, &raw_ann)?;
-            let pi_ann = normalize(tc_env, &lam_ann)?;
+            let pi_ann = nf_term(tc_env, &lam_ann)?;
             let (lam_body, pi_body) = {
                 let mut body_tc_env = tc_env.clone();
                 body_tc_env
@@ -751,7 +536,7 @@ pub fn infer_term(tc_env: &TcEnv, raw_term: &raw::RcTerm) -> Result<(RcTerm, RcT
                     let ((Binder(free_var), Embed(ann)), body) = scope.clone().unbind();
 
                     let arg = check_term(tc_env, raw_arg, &ann)?;
-                    let body = normalize(tc_env, &body.substs(&[(free_var, arg.clone())]))?;
+                    let body = nf_term(tc_env, &body.substs(&[(free_var, arg.clone())]))?;
 
                     Ok((RcTerm::from(Term::App(head, arg)), body))
                 },
@@ -773,7 +558,7 @@ pub fn infer_term(tc_env: &TcEnv, raw_term: &raw::RcTerm) -> Result<(RcTerm, RcT
 
             let (ann, ann_level) = infer_universe(tc_env, &raw_ann)?;
             let (body, body_level) = {
-                let ann = normalize(tc_env, &ann)?;
+                let ann = nf_term(tc_env, &ann)?;
                 let mut body_tc_env = tc_env.clone();
                 body_tc_env.declarations.insert(free_var.clone(), ann);
                 infer_universe(&body_tc_env, &raw_body)?
@@ -826,7 +611,7 @@ pub fn infer_term(tc_env: &TcEnv, raw_term: &raw::RcTerm) -> Result<(RcTerm, RcT
             match field_ty {
                 Some(field_ty) => Ok((
                     RcTerm::from(Term::Proj(expr, label.clone())),
-                    normalize(tc_env, &field_ty)?,
+                    nf_term(tc_env, &field_ty)?,
                 )),
                 None => Err(TypeError::NoFieldInType {
                     label_span,
