+++ src/semantics/tests/infer_term.rs
@@ -56,7 +56,7 @@ fn ty() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -70,7 +70,7 @@ fn ty_levels() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -84,7 +84,7 @@ fn ann_ty_id() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -98,7 +98,7 @@ fn ann_arrow_ty_id() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -125,7 +125,7 @@ fn app() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -156,7 +156,7 @@ fn lam() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -170,7 +170,7 @@ fn pi() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -184,7 +184,7 @@ fn id() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -198,7 +198,7 @@ fn id_ann() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -214,7 +214,7 @@ fn id_app_ty() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -229,7 +229,7 @@ fn id_app_ty_ty() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -243,7 +243,7 @@ fn id_app_ty_arr_ty() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -257,7 +257,7 @@ fn id_app_arr_pi_ty() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -271,7 +271,7 @@ fn apply() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -285,7 +285,7 @@ fn const_() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -299,7 +299,7 @@ fn const_flipped() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -313,7 +313,7 @@ fn flip() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -327,7 +327,7 @@ fn compose() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -345,7 +345,7 @@ fn case_expr() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -362,7 +362,7 @@ fn case_expr_bool() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -396,7 +396,7 @@ fn case_expr_wildcard() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -426,7 +426,7 @@ mod church_encodings {
 
         assert_term_eq!(
             parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-            parse_normalize(&mut codemap, &tc_env, expected_ty),
+            parse_nf_term(&mut codemap, &tc_env, expected_ty),
         );
     }
 
@@ -446,7 +446,7 @@ mod church_encodings {
 
         assert_term_eq!(
             parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-            parse_normalize(&mut codemap, &tc_env, expected_ty),
+            parse_nf_term(&mut codemap, &tc_env, expected_ty),
         );
     }
 
@@ -466,7 +466,7 @@ mod church_encodings {
 
         assert_term_eq!(
             parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-            parse_normalize(&mut codemap, &tc_env, expected_ty),
+            parse_nf_term(&mut codemap, &tc_env, expected_ty),
         );
     }
 
@@ -485,7 +485,7 @@ mod church_encodings {
 
         assert_term_eq!(
             parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-            parse_normalize(&mut codemap, &tc_env, expected_ty),
+            parse_nf_term(&mut codemap, &tc_env, expected_ty),
         );
     }
 }
@@ -500,7 +500,7 @@ fn empty_record_ty() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -514,7 +514,7 @@ fn empty_record() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -528,7 +528,7 @@ fn dependent_record_ty() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -555,7 +555,7 @@ fn proj() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -590,7 +590,7 @@ fn proj_weird1() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
 
@@ -611,7 +611,7 @@ fn proj_weird2() {
 
     assert_term_eq!(
         parse_infer_term(&mut codemap, &tc_env, given_expr).1,
-        parse_normalize(&mut codemap, &tc_env, expected_ty),
+        parse_nf_term(&mut codemap, &tc_env, expected_ty),
     );
 }
