+++ src/semantics/tests/normalize.rs
@@ -1,249 +1,254 @@
-use goldenfile::Mint;
-
-use std::io::Write;
-
 use super::*;
 
-fn golden(filename: &str, literal: &str) {
-    let mut codemap = CodeMap::new();
-    let tc_env = TcEnv::default();
+mod nf_term {
+    use syntax::core::{Head, Neutral, RcNeutral};
 
-    let path = "src/semantics/tests/goldenfiles";
+    use super::*;
 
-    let mut mint = Mint::new(path);
-    let mut file = mint.new_goldenfile(filename).unwrap();
+    fn golden(filename: &str, literal: &str) {
+        use goldenfile::Mint;
+        use std::io::Write;
 
-    let term = parse_normalize(&mut codemap, &tc_env, literal);
+        let mut codemap = CodeMap::new();
+        let tc_env = TcEnv::default();
 
-    write!(file, "{:#?}", term).unwrap();
-}
+        let path = "src/semantics/tests/goldenfiles";
 
-#[test]
-fn var() {
-    let tc_env = TcEnv::default();
+        let mut mint = Mint::new(path);
+        let mut file = mint.new_goldenfile(filename).unwrap();
 
-    let x = FreeVar::fresh_named("x");
-    let var = RcTerm::from(Term::Var(Var::Free(x.clone())));
+        let term = parse_nf_term(&mut codemap, &tc_env, literal);
 
-    assert_eq!(
-        normalize(&tc_env, &var).unwrap(),
-        RcValue::from(Value::from(Var::Free(x))),
-    );
-}
+        write!(file, "{:#?}", term).unwrap();
+    }
 
-#[test]
-fn ty() {
-    golden("ty", r"Type");
-}
+    #[test]
+    fn var() {
+        let tc_env = TcEnv::default();
 
-#[test]
-fn lam() {
-    let mut codemap = CodeMap::new();
-    let tc_env = TcEnv::default();
+        let x = FreeVar::fresh_named("x");
+        let var = RcTerm::from(Term::Var(Var::Free(x.clone())));
 
-    let x = FreeVar::fresh_named("x");
-
-    assert_term_eq!(
-        parse_normalize(&mut codemap, &tc_env, r"\x : Type => x"),
-        RcValue::from(Value::Lam(Scope::new(
-            (Binder(x.clone()), Embed(RcValue::from(Value::universe(0)))),
+        assert_eq!(
+            nf_term(&tc_env, &var).unwrap(),
             RcValue::from(Value::from(Var::Free(x))),
-        ))),
-    );
-}
+        );
+    }
 
-#[test]
-fn pi() {
-    let mut codemap = CodeMap::new();
-    let tc_env = TcEnv::default();
+    #[test]
+    fn ty() {
+        golden("ty", r"Type");
+    }
 
-    let x = FreeVar::fresh_named("x");
+    #[test]
+    fn lam() {
+        let mut codemap = CodeMap::new();
+        let tc_env = TcEnv::default();
 
-    assert_term_eq!(
-        parse_normalize(&mut codemap, &tc_env, r"(x : Type) -> x"),
-        RcValue::from(Value::Pi(Scope::new(
-            (Binder(x.clone()), Embed(RcValue::from(Value::universe(0)))),
-            RcValue::from(Value::from(Var::Free(x))),
-        ))),
-    );
-}
+        let x = FreeVar::fresh_named("x");
 
-#[test]
-fn lam_app() {
-    let mut codemap = CodeMap::new();
-    let tc_env = TcEnv::default();
-
-    let given_expr = r"\(x : Type -> Type) (y : Type) => x y";
-
-    let x = FreeVar::fresh_named("x");
-    let y = FreeVar::fresh_named("y");
-    let ty_arr = RcValue::from(Value::Pi(Scope::new(
-        (
-            Binder(FreeVar::fresh_unnamed()),
-            Embed(RcValue::from(Value::universe(0))),
-        ),
-        RcValue::from(Value::universe(0)),
-    )));
-
-    assert_term_eq!(
-        parse_normalize(&mut codemap, &tc_env, given_expr,),
-        RcValue::from(Value::Lam(Scope::new(
-            (Binder(x.clone()), Embed(ty_arr)),
+        assert_term_eq!(
+            parse_nf_term(&mut codemap, &tc_env, r"\x : Type => x"),
             RcValue::from(Value::Lam(Scope::new(
-                (Binder(y.clone()), Embed(RcValue::from(Value::universe(0)))),
-                RcValue::from(Value::Neutral(
-                    RcNeutral::from(Neutral::Head(Head::Var(Var::Free(x)))),
-                    vector![RcValue::from(Value::from(Var::Free(y)))],
-                )),
+                (Binder(x.clone()), Embed(RcValue::from(Value::universe(0)))),
+                RcValue::from(Value::from(Var::Free(x))),
             ))),
-        ))),
-    );
-}
+        );
+    }
 
-#[test]
-fn pi_app() {
-    let mut codemap = CodeMap::new();
-    let tc_env = TcEnv::default();
+    #[test]
+    fn pi() {
+        let mut codemap = CodeMap::new();
+        let tc_env = TcEnv::default();
 
-    let given_expr = r"(x : Type -> Type) -> (y : Type) -> x y";
+        let x = FreeVar::fresh_named("x");
 
-    let x = FreeVar::fresh_named("x");
-    let y = FreeVar::fresh_named("y");
-    let ty_arr = RcValue::from(Value::Pi(Scope::new(
-        (
-            Binder(FreeVar::fresh_unnamed()),
-            Embed(RcValue::from(Value::universe(0))),
-        ),
-        RcValue::from(Value::universe(0)),
-    )));
-
-    assert_term_eq!(
-        parse_normalize(&mut codemap, &tc_env, given_expr),
-        RcValue::from(Value::Pi(Scope::new(
-            (Binder(x.clone()), Embed(ty_arr)),
+        assert_term_eq!(
+            parse_nf_term(&mut codemap, &tc_env, r"(x : Type) -> x"),
             RcValue::from(Value::Pi(Scope::new(
-                (Binder(y.clone()), Embed(RcValue::from(Value::universe(0)))),
-                RcValue::from(Value::Neutral(
-                    RcNeutral::from(Neutral::Head(Head::Var(Var::Free(x)))),
-                    vector![RcValue::from(Value::from(Var::Free(y)))],
-                )),
+                (Binder(x.clone()), Embed(RcValue::from(Value::universe(0)))),
+                RcValue::from(Value::from(Var::Free(x))),
             ))),
-        ))),
-    );
-}
+        );
+    }
 
-// Passing `Type` to the polymorphic identity function should yield the type
-// identity function
-#[test]
-fn id_app_ty() {
-    let mut codemap = CodeMap::new();
-    let tc_env = TcEnv::default();
+    #[test]
+    fn lam_app() {
+        let mut codemap = CodeMap::new();
+        let tc_env = TcEnv::default();
 
-    let given_expr = r"(\(a : Type 1) (x : a) => x) Type";
-    let expected_expr = r"\x : Type => x";
+        let given_expr = r"\(x : Type -> Type) (y : Type) => x y";
 
-    assert_term_eq!(
-        parse_normalize(&mut codemap, &tc_env, given_expr),
-        parse_normalize(&mut codemap, &tc_env, expected_expr),
-    );
-}
+        let x = FreeVar::fresh_named("x");
+        let y = FreeVar::fresh_named("y");
+        let ty_arr = RcValue::from(Value::Pi(Scope::new(
+            (
+                Binder(FreeVar::fresh_unnamed()),
+                Embed(RcValue::from(Value::universe(0))),
+            ),
+            RcValue::from(Value::universe(0)),
+        )));
 
-// Passing `Type` to the `Type` identity function should yield `Type`
-#[test]
-fn id_app_ty_ty() {
-    let mut codemap = CodeMap::new();
-    let tc_env = TcEnv::default();
+        assert_term_eq!(
+            parse_nf_term(&mut codemap, &tc_env, given_expr,),
+            RcValue::from(Value::Lam(Scope::new(
+                (Binder(x.clone()), Embed(ty_arr)),
+                RcValue::from(Value::Lam(Scope::new(
+                    (Binder(y.clone()), Embed(RcValue::from(Value::universe(0)))),
+                    RcValue::from(Value::Neutral(
+                        RcNeutral::from(Neutral::Head(Head::Var(Var::Free(x)))),
+                        vector![RcValue::from(Value::from(Var::Free(y)))],
+                    )),
+                ))),
+            ))),
+        );
+    }
 
-    let given_expr = r"(\(a : Type 2) (x : a) => x) (Type 1) Type";
-    let expected_expr = r"Type";
+    #[test]
+    fn pi_app() {
+        let mut codemap = CodeMap::new();
+        let tc_env = TcEnv::default();
 
-    assert_term_eq!(
-        parse_normalize(&mut codemap, &tc_env, given_expr),
-        parse_normalize(&mut codemap, &tc_env, expected_expr),
-    );
-}
+        let given_expr = r"(x : Type -> Type) -> (y : Type) -> x y";
 
-// Passing `Type -> Type` to the `Type` identity function should yield
-// `Type -> Type`
-#[test]
-fn id_app_ty_arr_ty() {
-    let mut codemap = CodeMap::new();
-    let tc_env = TcEnv::default();
+        let x = FreeVar::fresh_named("x");
+        let y = FreeVar::fresh_named("y");
+        let ty_arr = RcValue::from(Value::Pi(Scope::new(
+            (
+                Binder(FreeVar::fresh_unnamed()),
+                Embed(RcValue::from(Value::universe(0))),
+            ),
+            RcValue::from(Value::universe(0)),
+        )));
 
-    let given_expr = r"(\(a : Type 2) (x : a) => x) (Type 1) (Type -> Type)";
-    let expected_expr = r"Type -> Type";
+        assert_term_eq!(
+            parse_nf_term(&mut codemap, &tc_env, given_expr),
+            RcValue::from(Value::Pi(Scope::new(
+                (Binder(x.clone()), Embed(ty_arr)),
+                RcValue::from(Value::Pi(Scope::new(
+                    (Binder(y.clone()), Embed(RcValue::from(Value::universe(0)))),
+                    RcValue::from(Value::Neutral(
+                        RcNeutral::from(Neutral::Head(Head::Var(Var::Free(x)))),
+                        vector![RcValue::from(Value::from(Var::Free(y)))],
+                    )),
+                ))),
+            ))),
+        );
+    }
 
-    assert_term_eq!(
-        parse_normalize(&mut codemap, &tc_env, given_expr),
-        parse_normalize(&mut codemap, &tc_env, expected_expr),
-    );
-}
+    // Passing `Type` to the polymorphic identity function should yield the type
+    // identity function
+    #[test]
+    fn id_app_ty() {
+        let mut codemap = CodeMap::new();
+        let tc_env = TcEnv::default();
 
-// Passing the id function to itself should yield the id function
-#[test]
-fn id_app_id() {
-    let mut codemap = CodeMap::new();
-    let tc_env = TcEnv::default();
+        let given_expr = r"(\(a : Type 1) (x : a) => x) Type";
+        let expected_expr = r"\x : Type => x";
 
-    let given_expr = r"
-        (\(a : Type 1) (x : a) => x)
-            ((a : Type) -> a -> a)
-            (\(a : Type) (x : a) => x)
-    ";
-    let expected_expr = r"\(a : Type) (x : a) => x";
+        assert_term_eq!(
+            parse_nf_term(&mut codemap, &tc_env, given_expr),
+            parse_nf_term(&mut codemap, &tc_env, expected_expr),
+        );
+    }
 
-    assert_term_eq!(
-        parse_normalize(&mut codemap, &tc_env, given_expr),
-        parse_normalize(&mut codemap, &tc_env, expected_expr),
-    );
-}
+    // Passing `Type` to the `Type` identity function should yield `Type`
+    #[test]
+    fn id_app_ty_ty() {
+        let mut codemap = CodeMap::new();
+        let tc_env = TcEnv::default();
 
-// Passing the id function to the 'const' combinator should yield a
-// function that always returns the id function
-#[test]
-fn const_app_id_ty() {
-    let mut codemap = CodeMap::new();
-    let tc_env = TcEnv::default();
+        let given_expr = r"(\(a : Type 2) (x : a) => x) (Type 1) Type";
+        let expected_expr = r"Type";
 
-    let given_expr = r"
-        (\(a : Type 1) (b : Type 2) (x : a) (y : b) => x)
-            ((a : Type) -> a -> a)
-            (Type 1)
-            (\(a : Type) (x : a) => x)
-            Type
-    ";
-    let expected_expr = r"\(a : Type) (x : a) => x";
+        assert_term_eq!(
+            parse_nf_term(&mut codemap, &tc_env, given_expr),
+            parse_nf_term(&mut codemap, &tc_env, expected_expr),
+        );
+    }
 
-    assert_term_eq!(
-        parse_normalize(&mut codemap, &tc_env, given_expr),
-        parse_normalize(&mut codemap, &tc_env, expected_expr),
-    );
-}
+    // Passing `Type -> Type` to the `Type` identity function should yield
+    // `Type -> Type`
+    #[test]
+    fn id_app_ty_arr_ty() {
+        let mut codemap = CodeMap::new();
+        let tc_env = TcEnv::default();
 
-#[test]
-fn horrifying_app_1() {
-    let mut codemap = CodeMap::new();
-    let tc_env = TcEnv::default();
+        let given_expr = r"(\(a : Type 2) (x : a) => x) (Type 1) (Type -> Type)";
+        let expected_expr = r"Type -> Type";
 
-    let given_expr = r"(\(t : Type) (f : (a : Type) -> Type) => f t) String (\(a : Type) => a)";
-    let expected_expr = r"String";
+        assert_term_eq!(
+            parse_nf_term(&mut codemap, &tc_env, given_expr),
+            parse_nf_term(&mut codemap, &tc_env, expected_expr),
+        );
+    }
 
-    assert_term_eq!(
-        parse_normalize(&mut codemap, &tc_env, given_expr),
-        parse_normalize(&mut codemap, &tc_env, expected_expr),
-    );
-}
+    // Passing the id function to itself should yield the id function
+    #[test]
+    fn id_app_id() {
+        let mut codemap = CodeMap::new();
+        let tc_env = TcEnv::default();
 
-#[test]
-fn horrifying_app_2() {
-    let mut codemap = CodeMap::new();
-    let tc_env = TcEnv::default();
+        let given_expr = r"
+            (\(a : Type 1) (x : a) => x)
+                ((a : Type) -> a -> a)
+                (\(a : Type) (x : a) => x)
+        ";
+        let expected_expr = r"\(a : Type) (x : a) => x";
 
-    let given_expr = r#"(\(t: String) (f: String -> String) => f t) "hello""#;
-    let expected_expr = r#"\(f : String -> String) => f "hello""#;
+        assert_term_eq!(
+            parse_nf_term(&mut codemap, &tc_env, given_expr),
+            parse_nf_term(&mut codemap, &tc_env, expected_expr),
+        );
+    }
 
-    assert_term_eq!(
-        parse_normalize(&mut codemap, &tc_env, given_expr),
-        parse_normalize(&mut codemap, &tc_env, expected_expr),
-    );
+    // Passing the id function to the 'const' combinator should yield a
+    // function that always returns the id function
+    #[test]
+    fn const_app_id_ty() {
+        let mut codemap = CodeMap::new();
+        let tc_env = TcEnv::default();
+
+        let given_expr = r"
+            (\(a : Type 1) (b : Type 2) (x : a) (y : b) => x)
+                ((a : Type) -> a -> a)
+                (Type 1)
+                (\(a : Type) (x : a) => x)
+                Type
+        ";
+        let expected_expr = r"\(a : Type) (x : a) => x";
+
+        assert_term_eq!(
+            parse_nf_term(&mut codemap, &tc_env, given_expr),
+            parse_nf_term(&mut codemap, &tc_env, expected_expr),
+        );
+    }
+
+    #[test]
+    fn horrifying_app_1() {
+        let mut codemap = CodeMap::new();
+        let tc_env = TcEnv::default();
+
+        let given_expr = r"(\(t : Type) (f : (a : Type) -> Type) => f t) String (\(a : Type) => a)";
+        let expected_expr = r"String";
+
+        assert_term_eq!(
+            parse_nf_term(&mut codemap, &tc_env, given_expr),
+            parse_nf_term(&mut codemap, &tc_env, expected_expr),
+        );
+    }
+
+    #[test]
+    fn horrifying_app_2() {
+        let mut codemap = CodeMap::new();
+        let tc_env = TcEnv::default();
+
+        let given_expr = r#"(\(t: String) (f: String -> String) => f t) "hello""#;
+        let expected_expr = r#"\(f : String -> String) => f "hello""#;
+
+        assert_term_eq!(
+            parse_nf_term(&mut codemap, &tc_env, given_expr),
+            parse_nf_term(&mut codemap, &tc_env, expected_expr),
+        );
+    }
 }
